pyproject.toml
[tool.poetry]
name = "elk-fastapi-app-monitoring"
version = "2.0.2"
description = ""
authors = ["olahsymbo <olahsymbo@gmail.com>"]
packages = [{ include = "elk_fastapi_app_monitoring", from = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"
python-semantic-release = "^8.5.2"
sqlalchemy = "^2.0.24"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
alembic = "^1.13.1"
coverage = "^7.4.0"
pytest = "^7.4.4"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
branch = "main"
version_variable = "src/elasticsearch_logstash_kibana/__init__.py:__version__"
version_source = "tag"
commit_version_number = true # required for version_source = "tag"
commit_message = "{version} [skip ci]" # skip triggering ci pipelines for version commits
tag_commit = true
upload_to_pypi = false
upload_to_release = false
hvcs = "github" # gitlab is also supported (hvcs = host version control system)
changelog_file = "CHANGELOG.md"             # changelog file
dist_path = "dist/"                         # where to put dists 

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "feat",
    "fix",
    "perf",
    "style",
    "refactor",
    "test",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.mypy]
files = ['rdflib', 'test', 'devtools', 'examples']
python_version = "3.8"
warn_unused_configs = true
ignore_missing_imports = true
disallow_subclassing_any = false
warn_unreachable = true
warn_unused_ignores = true
no_implicit_optional = false
implicit_reexport = false

[tool.isort]
profile = "black"
py_version = 37
line_length = 88
src_paths= ["rdflib", "test", "devtools", "examples"]
supported_extensions = ["pyw", "pyi", "py"]
skip = [
    '.eggs',         # exclude a few common directories in the
    '.git',          # root of the project
    '.hg',
    '.mypy_cache',
    '.pytest_cache',
    '.tox',
    '.venv',
    '.var',
    '.github',
    '_build',
    'htmlcov',
    'benchmarks',
    'test_reports',
    'rdflib.egg-info',
    'buck-out',
    'build',
    'dist',
    'venv',
]

[tool.black]
line-length = "88"
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | \.var
    | \.github
    | _build
    | htmlcov
    | benchmarks
    | test_reports
    | rdflib.egg-info
    | buck-out
    | build
    | dist
    | venv
  )/
)
'''

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]
[tool.coverage.report]
fail_under = 100